requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-cohere>=0.0.1
chromadb>=0.3.0
cohere>=4.37
streamlit>=1.22.0
pdfplumber>=0.9.0
python-docx>=0.8.11
tiktoken>=0.4.0
python-dotenv>=1.0.0
langchain_chroma==0.2.3